Props
Changed props interface to use the one in models/. Also changed a variable name.
