+++ src/models/ResourceProps.ts
@@ -1,0 +1,9 @@
+export interface ResourceProps {
+    id: number;
+    author: string;
+    title: string;
+    description: string;
+    filename: string;
+}
+
+export default ResourceProps;
